OO-948: commit the updated repository entry first to free the lock and play with the lifecycle after

+++ src/main/java/org/olat/repository/RepositoryManager.java
@@ -769,11 +769,14 @@ public class RepositoryManager extends BasicManager {
 		return entry;
 	}
 	
-	private void updateLifeCycle(RepositoryEntry reloadedRe) {
-		LifeCycleManager lcManager = LifeCycleManager.createInstanceFor(reloadedRe);
-		if (lcManager.lookupLifeCycleEntry(RepositoryDeletionManager.SEND_DELETE_EMAIL_ACTION) != null) {
-			log.audit("Repository-Deletion: Remove from delete-list repositoryEntry=" + reloadedRe);
-			lcManager.deleteTimestampFor(RepositoryDeletionManager.SEND_DELETE_EMAIL_ACTION);
+	private void updateLifeCycle(RepositoryEntry reloadedRe, Date previousLastUsage) {
+		if(reloadedRe == null) return;
+		if(previousLastUsage == null || previousLastUsage.getTime() < (System.currentTimeMillis() - (60 * 60 * 1000))) {
+			LifeCycleManager lcManager = LifeCycleManager.createInstanceFor(reloadedRe);
+			if (lcManager.lookupLifeCycleEntry(RepositoryDeletionManager.SEND_DELETE_EMAIL_ACTION) != null) {
+				log.audit("Repository-Deletion: Remove from delete-list repositoryEntry=" + reloadedRe);
+				lcManager.deleteTimestampFor(RepositoryDeletionManager.SEND_DELETE_EMAIL_ACTION);
+			}
 		}
 	}
 
@@ -783,14 +786,16 @@ public class RepositoryManager extends BasicManager {
 	 */
 	public RepositoryEntry incrementLaunchCounter(RepositoryEntry re) {
 		RepositoryEntry reloadedRe = loadForUpdate(re);
-		if(reloadedRe == null) return null;//deleted
-
-		reloadedRe.setLaunchCounter(reloadedRe.getLaunchCounter() + 1);
-		reloadedRe.setLastUsage(new Date());
-		updateLifeCycle(reloadedRe);
-		
-		RepositoryEntry updatedRe = DBFactory.getInstance().getCurrentEntityManager().merge(reloadedRe);
-		DBFactory.getInstance().commit();
+		RepositoryEntry updatedRe = null;
+		Date previousLastUsage = null;
+		if(reloadedRe != null) {
+			reloadedRe.setLaunchCounter(reloadedRe.getLaunchCounter() + 1);
+			previousLastUsage = reloadedRe.getLastUsage();
+			reloadedRe.setLastUsage(new Date());
+			updatedRe = dbInstance.getCurrentEntityManager().merge(reloadedRe);
+		}
+		dbInstance.commit();
+		updateLifeCycle(reloadedRe, previousLastUsage);
 		return updatedRe;
 	}
 
@@ -800,13 +805,16 @@ public class RepositoryManager extends BasicManager {
 	 */
 	public RepositoryEntry incrementDownloadCounter( final RepositoryEntry re) {
 		RepositoryEntry reloadedRe = loadForUpdate(re);
-		if(reloadedRe == null) return null;//deleted
-
-		reloadedRe.setDownloadCounter(reloadedRe.getDownloadCounter() + 1);
-		reloadedRe.setLastUsage(new Date());
-		updateLifeCycle(reloadedRe);
-		RepositoryEntry updatedRe = DBFactory.getInstance().getCurrentEntityManager().merge(reloadedRe);
-		DBFactory.getInstance().commit();
+		RepositoryEntry updatedRe = null;
+		Date previousLastUsage = null;
+		if(reloadedRe != null) {
+			reloadedRe.setDownloadCounter(reloadedRe.getDownloadCounter() + 1);
+			previousLastUsage = reloadedRe.getLastUsage();
+			reloadedRe.setLastUsage(new Date());
+			updatedRe = dbInstance.getCurrentEntityManager().merge(reloadedRe);
+		}
+		dbInstance.commit();
+		updateLifeCycle(reloadedRe, previousLastUsage);
 		return updatedRe;
 	}
 
@@ -835,8 +843,8 @@ public class RepositoryManager extends BasicManager {
 		reloadedRe.setAccess(access);
 		reloadedRe.setMembersOnly(membersOnly);//fxdiff VCRP-1,2: access control of resources
 		
-		RepositoryEntry updatedRe = DBFactory.getInstance().getCurrentEntityManager().merge(reloadedRe);
-		DBFactory.getInstance().commit();
+		RepositoryEntry updatedRe = dbInstance.getCurrentEntityManager().merge(reloadedRe);
+		dbInstance.commit();
 		return updatedRe;
 	}
 
@@ -855,8 +863,8 @@ public class RepositoryManager extends BasicManager {
 		if(StringHelper.containsNonWhitespace(description)) {
 			reloadedRe.setDescription(description);
 		}
-		RepositoryEntry updatedRe = DBFactory.getInstance().getCurrentEntityManager().merge(reloadedRe);
-		DBFactory.getInstance().commit();
+		RepositoryEntry updatedRe = dbInstance.getCurrentEntityManager().merge(reloadedRe);
+		dbInstance.commit();
 		return updatedRe;
 	}
 	
@@ -894,10 +902,10 @@ public class RepositoryManager extends BasicManager {
 		
 		RepositoryEntry updatedRe = DBFactory.getInstance().getCurrentEntityManager().merge(reloadedRe);
 		if(cycleToDelete != null) {
-			DBFactory.getInstance().getCurrentEntityManager().remove(cycleToDelete);
+			dbInstance.getCurrentEntityManager().remove(cycleToDelete);
 		}
 		
-		DBFactory.getInstance().commit();
+		dbInstance.commit();
 		return updatedRe;
 	}
 	
@@ -937,8 +945,8 @@ public class RepositoryManager extends BasicManager {
 		reloadedRe.setCanReference(canReference);
 		reloadedRe.setCanLaunch(canLaunch);
 		reloadedRe.setCanDownload(canDownload);
-		RepositoryEntry updatedRe = DBFactory.getInstance().getCurrentEntityManager().merge(reloadedRe);
-		DBFactory.getInstance().commit();
+		RepositoryEntry updatedRe = dbInstance.getCurrentEntityManager().merge(reloadedRe);
+		dbInstance.commit();
 		return updatedRe;
 	}
